+++ frontends/qt/resources/MacOS/convert_icon.sh
@@ -1,3 +1,5 @@
+# SPDX-License-Identifier: Apache-2.0
+
 convert ../app_icon_source.png -resize 1024x icon_1024x1024x32.png
 convert ../app_icon_source.png -resize 512x icon_512x512x32.png
 convert ../app_icon_source.png -resize 256x icon_256x256x32.png