+++ frontends/web/src/api/bitbox02.ts
@@ -1,18 +1,4 @@
-/**
- * Copyright 2023-2024 Shift Crypto AG
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// SPDX-License-Identifier: Apache-2.0
 
 import type { TUnsubscribe } from '@/utils/transport-common';
 import type { SuccessResponse, FailResponse } from './response';
@@ -82,6 +68,7 @@ type VersionInfoCommon = {
   canBackupWithRecoveryWords: boolean;
   canCreate12Words: boolean;
   canBIP85: boolean;
+  canChangePassword: boolean;
 };
 
 export type VersionInfo = VersionInfoCommon & (
@@ -229,4 +216,8 @@ export const attestationCheckDone = (
 
 export const bluetoothToggleEnabled = (deviceID: string): Promise<SuccessResponse | FailResponse> => {
   return apiPost(`devices/bitbox02/${deviceID}/bluetooth/toggle-enabled`);
+};
+
+export const changeDevicePassword = (deviceID: string): Promise<SuccessResponse | FailResponse> => {
+  return apiPost(`devices/bitbox02/${deviceID}/change-password`);
 };