+++ frontends/web/src/components/dialog/dialog.tsx
@@ -1,19 +1,4 @@
-/**
- * Copyright 2018 Shift Devices AG
- * Copyright 2021-2025 Shift Crypto AG
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// SPDX-License-Identifier: Apache-2.0
 
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { CloseXDark, CloseXWhite } from '@/components/icon';
@@ -165,7 +150,7 @@ export const Dialog = ({
       className={overlayClass}
     >
       <UseBackButton handler={closeHandler} />
-      <div className={modalClass}>
+      <div className={modalClass} ref={contentRef}>
         {title && (
           <div className={headerClass}>
             <h3 className={style.title}>{title}</h3>
@@ -177,7 +162,7 @@ export const Dialog = ({
             )}
           </div>
         )}
-        <div className={contentClass} ref={contentRef}>
+        <div className={contentClass}>
           <div className={style.content}>{children}</div>
         </div>
       </div>