+++ frontends/web/src/hooks/keyboard.ts
@@ -1,18 +1,4 @@
-/**
- * Copyright 2022 Shift Crypto AG
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// SPDX-License-Identifier: Apache-2.0
 
 import { useCallback, useEffect, useRef } from 'react';
 
@@ -54,14 +40,7 @@ export const useEsc = (
   });
 };
 
-const FOCUSABLE_SELECTOR = `
-  a:not(:disabled),
-  button:not(:disabled),
-  input:not(:disabled),
-  select:not(:disabled),
-  textarea:not(:disabled),
-  [tabindex]:not([tabindex="-1"]):not(:disabled)
-`;
+const FOCUSABLE_SELECTOR = '[autofocus]:not(:disabled)';
 
 /**
  * Traps focus inside the given ref when active,