+++ frontends/web/src/routes/account/actionButtons.tsx
@@ -1,23 +1,9 @@
-/**
- * Copyright 2022-2024 Shift Crypto AG
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// SPDX-License-Identifier: Apache-2.0
 
 import { MouseEvent } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
-import { WalletConnectLight } from '@/components/icon';
+import { ArrowFloorDownWhite, ArrowFloorUpWhite, Coins, WalletConnectLight } from '@/components/icon';
 import { useMediaQuery } from '@/hooks/mediaquery';
 import { AccountCode, TAccount, CoinCode } from '@/api/account';
 import { isEthereumBased } from './utils';
@@ -68,6 +54,7 @@ export const ActionButtons = ({ canSend, code, coinCode, exchangeSupported, acco
         to={sendLink}
         onClick={isEthereumBased(coinCode) ? maybeRouteSend : undefined}
       >
+        <ArrowFloorUpWhite width={16} height={16} />
         <span>{t('generic.send')}</span>
       </ButtonLink>
 
@@ -77,6 +64,7 @@ export const ActionButtons = ({ canSend, code, coinCode, exchangeSupported, acco
         primary
         to={`/account/${code}/receive`}
       >
+        <ArrowFloorDownWhite width={16} height={16} />
         <span>{t('generic.receiveWithoutCoinCode')}</span>
       </ButtonLink>
 
@@ -87,6 +75,7 @@ export const ActionButtons = ({ canSend, code, coinCode, exchangeSupported, acco
           primary
           to={`/market/info/${code}`}
         >
+          <Coins width={17} height={17} />
           <span>{t('generic.buySell')}</span>
         </ButtonLink>
       )}