+++ frontends/web/src/routes/account/info/buy-receive-cta.tsx
@@ -1,18 +1,4 @@
-/**
- * Copyright 2023 Shift Crypto AG
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// SPDX-License-Identifier: Apache-2.0
 
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -23,7 +9,7 @@ import { Button } from '@/components/forms';
 import { Balances } from '@/routes/account/summary/accountssummary';
 import { isBitcoinCoin, isEthereumBased } from '@/routes/account/utils';
 import { getVendorSupportedAccounts } from '@/routes/market/utils';
-import { WalletConnectLight } from '@/components/icon';
+import { ArrowFloorDownWhite, Coins, WalletConnectLight } from '@/components/icon';
 import { useMountedRef } from '@/hooks/mount';
 import { SubTitle } from '@/components/title';
 import styles from './buy-receive-cta.module.css';
@@ -77,7 +63,8 @@ export const BuyReceiveCTA = ({
       </p>
       <div className={styles.buttons}>
         {balanceList && (
-          <Button primary onClick={onReceiveCTA}>
+          <Button className={styles.button} primary onClick={onReceiveCTA}>
+            <ArrowFloorDownWhite width={18} height={18} />
             {/* "Receive Bitcoin", "Receive crypto" or "Receive LTC" (via placeholder "Receive {{coinCode}}") */}
             {t('generic.receive', {
               context: isBitcoin ? 'bitcoin' : (unit ? '' : 'crypto'),
@@ -86,7 +73,8 @@ export const BuyReceiveCTA = ({
           </Button>
         )}
         {(exchangeSupported && !isMobile) && (
-          <Button primary onClick={onMarketCTA}>
+          <Button className={styles.button} primary onClick={onMarketCTA}>
+            <Coins width={18} height={18} />
             {t('generic.buySell')}
           </Button>
         )}