+++ frontends/web/src/routes/settings/components/manage-accounts/dialogs/enableRememberWalletDialog.tsx
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: Apache-2.0
+
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Dialog, DialogButtons } from '@/components/dialog/dialog';